SpatialAudio: Handle case where there are no sources yet
It is possible that when filling the audio output buffer with data from
Resonance Audio with FillInterleavedOuputBuffer that it will return
false when no input source have been registered yet. This isn't
necessarily an error condition that requires the audio processing thread
to stall if there are no audio sources provided, since in that case the
correct thing to do if we expect there will be sources soon is to fill
the buffer with silence.

The problem with bailing out too soon is that when the sources are
finally loaded then the audio thread has stopped. This issue was
happening on the Meta Quest when in XR mode, likely due more happening
on the main thread currently leading to a longer time between when the
audio thread starts vs when the spatial sounds are loaded and
registered.

Change-Id: I753c64b968cd3223c690fb62f6e4deebb89590ea
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>