Make MockMultimediaPlugin an optional dependency
The MockMultimediaPlugin is a static plugin that
should only be used when it is an explicit
dependency, and should not be linked into any
other applications. To achieve this, it needs to
have the DEFAULT_IF set to FALSE. Without this
flag, all apps compiled against Qt will get an
automatic dependency on the mock plugin, and
linking fails because the plugin is not located in
the default location.

Amends 5070a7343504bbde988a903c2ef0a2d18af4303b

Pick-to: 6.7 6.6 6.5
Change-Id: Ic506420ba265a1d63eca9e1dab666df0fe480832
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>