Android-ffmpeg: Additional validation for video encoders
It appears that ffmpeg lib returns a list of encoders that may not be
available on the device. For this reason, avcodec_open2 may not work
if the codec is incorrectly selected.

To avoid this situation on Android, the list of available encoders is
prepared using the Android API ( [0]MediaCodecList ).

Unfortunately, [1]isHardwareAccelerated method was added in API 29, so
we are not allowed to use it. Instead, the isSoftwareCodec() method has
been added (taken from Android[2] sources).

[0] https://developer.android.com/reference/android/media/MediaCodecList
[1] https://developer.android.com/reference/android/media/MediaCodecInfo#isHardwareAccelerated()
[2] https://android.googlesource.com/platform/frameworks/av/+/master/media/libstagefright/MediaCodecList.cpp#320

Pick-to: 6.5
Fixes: QTBUG-121221
Change-Id: I4fddbe1043cf19a677ac3ed1ec21f8930358c431
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 4473084fe19dac3398bf230e0a189ee9645161ef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit ecb82b2f5b44d530818522d611e922304a67a449)