Rework QPulseAudioSink underflow error handling
The way PulseAudio sink was notifying about underflow errors was too
unreliable and prone to bugs. It relied on a m_resuming flag being set
correctly in multiple parts of the code, as well as the tick timer being
running even in push mode just to set the error afterwards. Additionally
the state was reset in the stream drain callback causing flakiness.

The error state is now set in the assigned underflow callback, and the
state is not updated in the drain callback. The timer is refactored for
pull mode use only.

Pick-to: 6.6 6.5
Change-Id: I9489ca3118c11c70cd1f77a8e5b93a04ef60dca8
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
(cherry picked from commit 3ad304b78c0e5f84ce78790e74a1c9da054fa304)
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>