tst_qmediaplayerbackend: Improve test to catch errors in looped playback
The previous test failed to detect an error that stopped the signal
QMediaPlayer::positionChanged to be emitted during looped playback.
This is now solved by checking that positionChanged is emitted
more than a certain amount of times during each loop iteration.

The positionChangingIntervals function is replaced by loopIterations()
which creates a vector of LoopIteration structs containing start- and
end position and the number of position changes per video loop
iteration. The only practical change here is that we now check for
an expected number of position changes per iteration.

In addition, the fixture pattern has been taken in use for finiteLoops()

Task-number: QTBUG-122045
Pick-to: 6.7 6.6 6.5
Change-Id: Ie15b81a1b3733835e7c3e20f8e514dbf4765321d
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>