tst_qaudiosink: Blacklist pullResumeFromUnderrun on RHEL 9.2
We want to be able to utilize the underflow callback of QPulseAudioSink
better. This has an effect on the pullResumeFromUnderrun test case.

The underflow callback in QPulseAudioSink is not correctly called when
using PipeWire, and RHEL 9.2 uses PipeWire by default. Therefore we
should blacklist the test case until it's fixed.

Task-number: QTBUG-122309
Pick-to: 6.7 6.6 6.5
Change-Id: I84656475528de082ece1e52077a467117777a5ac
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>