Fix "Media Player Example" deployment to Android
Changes in androiddeployqt change the way plugins are deployed to
an Android APK. Mainly, when using CMake, they are only deployed
if the module is explicitly added to the project.

A quick scan revealed that this example would not deploy under the new
changes, and a quick test revealed that it doesn't. Let's fix that.

Fixes: QTBUG-122640
Pick-to: 6.6 6.5
Change-Id: I1fad9925095b98dca1a70a3bfff8bd624cbf8a0a
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 1e7adcc64b3b17618e833764294759b49bae8ea1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>