Clarify documentation for QMediaMetaData::Key::ThumbnailImage
QMediaMetaData will only return a QImage for the key
QMediaMetaData::Key::ThumbnailImage if the media contains an embedded
thumbnail image. Specifically, Qt Multimedia will not generate an image
if it is not already part of the metadata on the media.

This is already documented in the Common attributes, but this patch
makes it more explicit, also for the QMediaMetaData::Key::ThumbnailImage
doc string.

Task-number: QTBUG-116324
Pick-to: 6.7 6.6 6.5
Change-Id: I537c7349aa08ecaf72f0a3b449744a6261f975aa
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim@klingt.org>