Android-FFMpeg: Dynamically create/remove AndroidSurfaceTexture
Before this change, MediaCodecTextureConverter used a single static
AndroidSurfaceTexture. When trying to play another video, currently used
Surface was trying to be reused. Because of that, more than one
simultaneous video playback caused the app to crash on Android.

Android surface textures need to be created dynamically, depending as
needed.

MediaCodecTextureConverter is created as a backend in TextureConverter
when the first frame from videoSink appears. At this point the Surface
is already created, set in the codecContext and used by FFmpeg.

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-121182
Fixes: QTBUG-122649
Change-Id: Ie9845ac78d770e4c06b6046fc59c0afbd39fb14a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>