Play entire sound with resampled data in QSoundEffect
If FFmpeg resampler is used to convert an audio signal from mono to
stereo in QSoundEffect, the number of bytes of data doubles. This patch
makes sure that the byte count of the resampled buffer is used instead
of the number of bytes in the raw sample buffer. Before this patch,
QSoundEffect could only play the first half of a WAV file if the file
contained mono signals.

This change makes sense for both the QSoundEffectPrivate size() and
bytesAvailable() functions because QSoundEffectPrivate can only read
data from the resampled audio buffer, never from the raw sample buffer.

Amends: f0016c4763361292c3eaf153e184b9b110f44c61

Fixes: QTBUG-122750
Pick-to: 6.7 6.6 6.5
Change-Id: Icbfa6f5c4b5f85d02dca875e76f01a046ce49143
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>