windows: fix video playback when audio output is not specified
When setting up the topology in MFPlayerSession, we add source and
output nodes for each source stream. But in case when output node can't
be created for some reason, e.g. this is an audio node and audio output
is not specified (m_audioOutput == nullptr), we may result in an error
on setting the topology (IMFMediaSession::SetTopology), like "A stream
descriptor was set on a source stream node but it was not selected on
the presentation descriptor."

Therefore it was causing an issue where the video would not play at all. To fix this, we need to simply remove the source node from the topology
if the output node cannot be created.

Fixes: QTBUG-122608
Fixes: QTBUG-122817
Pick-to: 6.6 6.5
Change-Id: I046a78a257890b1d1c182f5f1e3dc48761985afa
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit b59bf3fd57c4cebd853ac96a505c1b249c00e5f5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>