Simplify lifetime management of QPlatformAudioResampler
QPlatformMediaIntegration creates various objects, and return them as
QMaybe<Type*>. This moves responsibility of ownership to the caller.

This patch makes the createAudioResampler function return the resampler
as QMaybe<std::unique_ptr<Type>> which simplifies lifetime management.

This introduces an asymmetry with other factory functions, but this is
a step in the right direction

Pick-to: 6.7 6.6 6.5
Change-Id: Id2f67597c8d8271b184321d138e126edff5f3202
Reviewed-by: Tim Blechmann <tim@klingt.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>