FFmpeg EGL/vaapi: Disable HW accelerated texture conversion when failing
This change applies to hardware accelerated texture conversion using EGL
and VA-API. If a call to eglCreateImage() inside
VAAPITextureConverter::getTextures() fails with EGL_BAD_MATCH error,
this change makes sure we fall back to software decoding by setting
this->rhi = nullptr.

This specific error can be reproduced in a Linux environment with both
a NVIDIA and AMD gpu, when NVIDIA is used as the EGL vendor.

By initializing the hw context manually using vaInitialize(), it is
possible to detect this problem before trying to decode frames. But we
can't since we let FFmpeg do this for us when we call
av_hwdevice_ctx_create().

The related qWarning() messages are also elaborated with error codes
and more relevant information.

Fixes: QTBUG-112312
Pick-to: 6.7 6.6 6.5
Change-Id: Ie90623c7b2b18b70b1000ae3ad6f96872933e884
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>