FFmpeg Encoder: Do not proceed if AVIOContext is not set
According [0]documentation, when passing AVFormatContext to
avformat_write_header function, "Its pb field must be set to an already
opened AVIOContext". It seems that if pb is not set, calling
avformat_write_header will casue the application crash.

That is why we should not proceed if AVIOContext is not set.

[0]https://ffmpeg.org/doxygen/5.0/group__lavf__encoding.html#ga18b7b10bb5b94c4842de18166bc677cb

Task-number: QTBUG-122224
Pick-to: 6.7 6.6 6.5
Change-Id: I7d3ca40459bf8e77cf664d3b20f0c124a6283a06
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>