FFmpeg: Remove a false positive assertion from demuxer constructor
FFmpeg demuxer constructor assumed that the current loop index is less
than total loop count. This overlooked the fact that the demuxer is
recreated when seeking, which caused the assertion to fail for example
when the loop counter was reset to 1 before seeking.

This patch removes this assertion from the constructor.

Fixes: QTBUG-123117
Pick-to: 6.7 6.6 6.5
Change-Id: If7b4c367fb549540fb55cb50de10d70acdd0009d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>