Fix size of QVideoSink when the video is rotated
The size of QVideoSink impacts on sizeHint of QVideoWidget,
and it can be taken by users for their own size hints.

All renderers draw rotated videos like this: if the frame size is
AxB, and the rotation is 90 degrees, the rendering size becomes BxA.

Let's keep QVideoSink::videoSize consistent.

Added a function to qtmm utilities to get rotates size.

Pick-to: 6.7 6.6 6.5
Change-Id: Ia0cb6f43a598820007fac8178104631a80d96dfb
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>