QMediaStorageLocation: Do not modified path with content scheme
QMediaStorageLocation::generateFileName function is trying to create
whole path with file name according to the provided parameters.

In case when requestedName is an URI content:// scheme (used on Android)
we should not modified it at all. The structure of content URIs is
determined by the content provider implementation. Modifying it may lead
to errors or unintended behavior in application.

Additionally, instead of QStringLiteral start to use modern _L1 string
literal operator.

Pick-to: 6.7 6.6 6.5
Task-number: QTBUG-122224
Change-Id: Ia888a49361da3d2d87365551097d763413350f14
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>