Fix keeping frame mapped after the target OpenGL texture released
It's a regression after
codereview.qt-project.org/c/qt/qtmultimedia/+/537062
The added test doesn't work without the fix.

The main reason of the regression is that the OpenGl rendering
pipeline in QRhi keeps the source QImage/QByteArray for
some time after the texture loading and even the texture releasing.

It's not going to be fixed in QRhi now, the most relevant approach
seem to be fixing on QtMM side with matching tests.

We're choosing the approach suggested in
codereview.qt-project.org/c/qt/qtmultimedia/+/541083

Pick-to: 6.7 6.6 6.5
Change-Id: I1822a5f61c022dc6375e41a984f13b23137b9b5a
Reviewed-by: Tim Blechmann <tim@klingt.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>