FFmpeg: Hardware-accelerated codec validation
In 4473084fe19dac3398bf230e0a189ee9645161ef commit an additional
validation for video hardware-accelerated codecs was added (in Android).
Android API is used to get list of codecs available on device.

It seems that in case when pix_fmts (array of supported pixel formats)
is empty we can still accept unavailable hardware-accelerated codec.

The main problem here is that there is no API in ffmpeg to check if
codec is used for hardware or software acceleration. There is only the
AV_CODEC_CAP_HARDWARE flag that indicates that a particular codec
supports hardware acceleration (codec is backed by a hardware
implementation).

It seems that Android for hardware acceleration uses the "MediaCodec"
system library (while for software uses ffmpeg lib implementation). That
is why we assume that mediacodec will be use for hardware acceleration
and should be available on device.

Fixes: QTBUG-122575
Pick-to: 6.7 6.6 6.5
Change-Id: Ifbb19a2ae8ec01c7a06d171a5a7eceefa14c1355
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>