Android-ffmpeg: Do not keep crop region for Camera Zoom
Instead of keeping the crop area scaled as a rectangle, just keep the
zoomFactor. Calculate scaler crop region when needed.

In case when switch the camera, value of
[0]SENSOR_INFO_ACTIVE_ARRAY_SIZE may change. In this case, we should not
use the previously calculated Zoom values.

This change also allows to update internal zoom value when camera is not
started.

[0]https://developer.android.com/reference/android/hardware/camera2/CameraCharacteristics#SENSOR_INFO_ACTIVE_ARRAY_SIZE

Pick-to: 6.7 6.6 6.5
Task-number: QTBUG-122968
Change-Id: Idbb201febf7e09a3540bb48e1521409190647474
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>