Use std::addressof to get the address of QMaybe wrapped type
Using QMaybe with types that overloads operator& can have unintended
effects because QMaybe::operator-> would call the overload. This could
cause problems with types such as WRL::ComPtr which resets its pointer
when taking its address.

To ensure that we get the address of the instance, we should use
std::addressof instead of calling operator& to get the wrapped object's
address.

This fixes an issue where QMaybe<ComPtr> resets the COM pointer and
returns nullptr when attempting to extract the raw interface pointer
like this:

    QMaybe<ComPtr<Interface>, HRESULT> instance = foo();
    Interface *rawIf = instance->Get(); // Reset instance and return 0

Pick-to: 6.7 6.6 6.5
Change-Id: I544dec6dc0cf59c2b3a46fe6d9ebbf08aac80737
Reviewed-by: Tim Blechmann <tim@klingt.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>