FFmpegVideoEncoder: Process frames also in pause state
Remove the pause flag check from processOne() method in
QVideoEncoder.

When pause flag is set, no more frames will be added to the queue in
addFrame method anyway. That is why additional check in processOne
method is not needed and we can allow to handle already delivered
frames.

This removal fixes a possible infinity loop that occurs when recording
is stopped in the pause state. This will happen when m_videoFrameQueue
is not empty. In the cleanup method we will try to process all frames,
but we will not process frames due to the pause state.

Pick-to: 6.6 6.5
Fixes: QTBUG-122184
Change-Id: I1cc307bb140b934d130bae08a59fd61f08111b84
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 653d196a1c77aba2f80566c114f4646af676f69e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>