FFmpeg: Check format validity when updating buffer size
QFFmpegAudioInput's internal IODevice gets the input buffer size using
the preferred format of the currently set audio device. When setting a
null audio input device, the buffer size was set to 0.

This created a possibility for an infinite loop when an audio source
wants to write recording stream data to the buffer.

This patch adds a check to setting the frame size of QFFmpegAudioInput
to ensure that the buffer size is set to default on invalid formats
instead of 0.

Task-number: QTBUG-123224
Task-number: QTBUG-123177
Pick-to: 6.6 6.5
Change-Id: I540cd6c6874e5efed98d9a52b67676e4439a96c5
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 644265165b586c285c5952acc8ad29a2204ed4fa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>