Android: Restart camera after QCameraFormat changes
Before this change, QCameraFormat was not immediately updated. It was
remembered and updated only after restarting the camera. Because of
this, the camera continued to generate frames with the old resolution.
When recording the video, each frame had to be scaled to obtain the
desired resolution.

Task-number: QTBUG-122140
Pick-to: 6.5
Change-Id: If30ef071f0140bfde4e08afe43bbb9e392dc6e82
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 562fc1ea0ae2feb76cba0ff55ed2acc92944a3f4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>