QAudioDecoder: fix unit tests for GStreamer
When setting a QUrl with an empty scheme, we need to set the file://
pattern for uridecodebin. This has been done for QMediaPlayer in the
past, but not for QAudioDecoder. Compare:
8a3806f6263f56b5be5b557e321b9b13d1913d5f

To keep consistency with QMediaPlayer, we introduce a
QAudioDecoderPrivate class.

Pick-to: 6.5
Task-number: QTBUG-96985
Change-Id: If954a04a003aecb0f862a7e5041ead43e4c9d259
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit d928939dc603d5819f960245b860e749b1f4e3a7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>