GStreamer: QGStreamerAudioSource fix potential memory leak
`invokeMethod` is not called if the object is destroyed before the slot
is invoked.
We replace the explicit reference counting with a handle type, which is
passed via a lambda function. This also allows us so remove the moc-ing

Pick-to: 6.6 6.5
Change-Id: I50eb3b23b5c83b01c98f6f3a99b0d5889bde718d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 4ef4fa1aa0181adaaaa33aa86286c3c4ff7d93af)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>