Android: Increase maximum number of images that can be acquired
Increase maximum number of images that can be acquired from 10 to 12 for
Android Camera.
The reason for this is the m_maxQueueSize value in VideoEncoder in the
ffmpeg-backend implementation. m_maxQueueSize limitation is 10 and at
the same time one additional frame can be processed. In such case we
still want to be able to get one more frame, that is why the
MaxNumberFrames is changed to 12

Task-number: QTBUG-122140
Pick-to: 6.7 6.5
Change-Id: I2b884d5dc7101f7bcc8794646e363a4410865819
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>