Android-backend: Do not reset QAndroidVideoOutput if not needed
Before this commit, QAndroidTextureVideoOutput was reset every time the
player (or camera preview) was stopped. Resetting
AndroidTextureVideoOutput clears the QRhi - that can cause a black video
frame problem (especially when trying to stop the current video and play
the next one).

This commit removes unnecessary QRhi cleanup. QAndroidVideoOutput still
need to be reset (QRhi need to be recreated) in case when
VideoSink::rhiChanged was emitted.

Fixes: QTBUG-119914
Fixes: QTBUG-117771
Pick-to: 6.7 6.5
Change-Id: Idc45bafe24cf7a8679b57555b0bbb5155979e11f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>