Recreate all Direct3D resources when FFmpeg device changed
When recalling a video file, FFmpeg recreates the underlying decoder
Direct3D device. All resources created on the previous device must
therefore be recreated.

Previously, we did not recreate all Direct3D resources if the opened
media file had the same size and format as previously loaded file.

This patch checks if the FFmpeg device has changed, and recreates
a temporary texture if needed.

Also, we add a check to recreate necessary resources if the RHI device
has changed.

Fixes: QTBUG-124253
Pick-to: 6.5 6.7
Change-Id: Ibe626c8dd6cdb8c05591a9b635e82647afa16042
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>