declarative-camera example: Hide video View button if none recorded
The View button was visible even when no video was recorded. This was
due to the assumption that actualLocation would evaluate to false if
empty (see QTBUG-63629).

This patch fixes the evaluation by converting the url to string, and
comparing to empty string.

Fixes: QTBUG-121172
Pick-to: 6.5
Change-Id: I642b03b439f57002bc505c46ee0045dc7cdbe514
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 71ed67972a929a4776cd6b781426a56679629e63)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>