MediaPlayer: Use palette, instead of hard coded color values
We want our examples to look good, independent of the system theme.
By using palette color values, instead of hard coded values, the palette
is able to change the colors in use, based on what theme the user has.

The example looked great when the user was using a mostly light colored
theme, but since the background of the PlaybackControl was hardcoded to
a whiteish color, there would barely be any contrast between the text of
buttons on top of it, and the background.

Fixes: QTBUG-124734
Pick-to: 6.10 6.8
Change-Id: I3b2157322e34347d947a442f37689f89ddd8225b
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>