GStreamer: AudioDecoder - fix race condition
The app sources lives on a gstreamer thread. `bufferAvailableChanged`
was called from both application thread and gstreamer thread, which is
not thread safe.

We simplify the application logic by deferring all mutation of the
`QGstreamerAudioDecoder` state to the application thread. This seems to
fix a spurious failure in tst_QAudioDecoderBackend::fileTest

Pick-to: 6.5 6.7
Change-Id: I992c39b071ebd48d5fc5ca28fae499c855b93304
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>