Fix mutex usage in encoders of the FFmpeg recorder
It was a mess of using data locking / conv var notifications.
The mutex in the ConsumerThread is supposed to protect
threading data (frames queue) and the exit flag.
The confitional variable must be invoked after releasing of
this mutex, but not after releasing a separate m_queueMutex
that was used before.

Pick-to: 6.7 6.5
Change-Id: I9010f4e9656b6298793b400211ccc82e12401e36
Reviewed-by: Tim Blechmann <tim@klingt.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>