FFmpeg/VAAPI: Enable finding libva shared libraries from cache
This change prevents an issue where Qt fails to load the VA-API shared
library libva, even though FFmpeg manages to find it. By adding the
version number of the latest libva.so file as an argument to the
overloaded function QLibrary::resolve, we make it possible for Qt
to find libva in the cache if FFmpeg has already loaded it.

Libva so files are numbered with VA-API version + 1, according to
the source files:
https://github.com/intel/libva/blob/master/configure.ac#L64
E.g. VA-API version 1.14 generates the filename libva.so.2

Pick-to: 6.7 6.5
Task-number: QTBUG-124586
Change-Id: I76c5a13d5f421272cff0ccfc613ec7cf0b148b06
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>