Simplify color management tests by using FFmpeg color conversion
This patch replaces custom conversion RGB to YUV conversion code used
only for tests with FFmpeg libswscale code. This makes it easier to
extend the tests with more pixel formats, and allow us to generate
realistic output images with correct color space and transfer.

The idea of this is that all reference images should now be equal, apart
from quantization errors from full vs limited range signals. In truth,
they are not exactly the same, and some color discrepancy remains.
Comments are added to the tests to address this.

Note that the maximum pixel diff threshold had to be adjusted because of
numerical inaccuracies on some platforms. Also, the reference image has
been reduced in size to reduce amount of space required for the
reference images. In addition, the image is now even width and height
because odd width/height images are not supported with chroma
subsampling.

Task-number: QTBUG-124537
Task-number: QTBUG-124534
Task-number: QTBUG-123749
Pick-to: 6.7 6.5
Change-Id: I0a36255de4acd75e61df0bb934d65ccff0ead063
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>