AVMediaPlayerObserver: deactivate audio session only if we activated it
It's possible to have several media players (imagine one player for
playing a background music in a game and the second providing an audio-feedback
for some user interaction). Setting a media source is documented
to stop the player's playback, but this does not mean _any_ media player
playing at the moment. AVAudioSession.sharedSession though acts as a
singleton and thus will stop all audio. To fix this, we only deactivate
audio session if we activated it previously (and since 'sharedInstance'
is a singleton anyway - alas not having 'active' property - we
add an activation count).

Pick-to: 6.7 6.5
Fixes: QTBUG-124298
Change-Id: Ia1dae7ad16d7affc64ab2b22517d5ad1d47ca379
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>