Android: Use correct Constant Value
STREAM_MUSIC[0] is not the correct value for setMode[1] function. It was
probably used by mistake.
MODE_IN_COMMUNICATION[2] is the correct one. This allows to properly
control audio devices (e.g. Bluetooth or speaker).

MODE_IN_COMMUNICATION has the same value (3) as STREAM_MUSIC, so nothing
changes in behavior. Just the correct name of constant value is used.

[0]https://developer.android.com/reference/android/media/AudioManager#STREAM_MUSIC
[1]https://developer.android.com/reference/android/media/AudioManager#setMode(int)
[2]https://developer.android.com/reference/android/media/AudioManager#MODE_IN_COMMUNICATION

Pick-to: 6.7 6.5
Change-Id: I1b198ff8a088f44b5bea5ab94f11a6f9cb2709de
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>