FFmpeg: Add QT_FFMPEG_PROTOCOL_WHITELIST environment variable
FFmpeg AV format context has a protocol whitelist option, which is
populated on avformat_open_input() call. This usually ends up correct,
but for example when opening an .sdp file for playing an rtp stream, the
whitelist doesn't include "rtp" and "udp".

This patch adds a check for the QT_FFMPEG_PROTOCOL_WHITELIST environment
variable, which the user can set to override the whitelist with their
own.

Fixes: QTBUG-125006
Pick-to: 6.7 6.5
Change-Id: I7ddb53f879af909fb8b38c639c571822542d6638
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>