Android: Use AudioDeviceCallback to track changes in audio devices
Instead of using BroadcastReceiver with different filters, use
AudioDeviceCallback[0] prepared for this purpose.

Since we are interested in changes in audio devices, we should use the
API prepared for this purpose. We shouldn't rely on BroadcastReceiver
- especially for a Bluetooth case. After receiving intent about the
connection of Bluetooth devices, AudioManager updates its list of
connected devices with a delay. Therefore, we cannot assume that the
device array will be updated immediately. That is why it is better to
use AudioDeviceCallback.

[0]https://developer.android.com/reference/android/media/AudioDeviceCallback

Task-number: QTBUG-123044
Pick-to: 6.5
Change-Id: Ic2087f536a045562386be6a5862c3ce7968210e7
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 815957bb0f281cafb471c6263c573e1bea7743b3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>