Android: Use MODE_NORMAL for speaker if possible
Using MODE_IN_COMMUNICATION allows to properly control audio devices
(like changes between Bluetooth and speaker), but the consequences are
the issues with Volume Button. At begginig it changes the "Call volume".
After some time it starts to control the "Media volume".

If there are no bluetooth or wired headset, the MODE_NORMAL will be
used for speaker. That allows to avoid issue with Volume Button.

In case when there are bluetooth or wired headset we need to use
MODE_IN_COMMUNICATION anyway.

Fixes: QTBUG-123044
Pick-to: 6.7 6.5
Change-Id: I7bfea28a971bca1b596622bf604f65941c0afe42
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>