tst_qmediaplayerbackend: Add test for playing RTP stream using .sdp file
Previous tests for stream playback only accounted for RTSP sessions. RTP
streams described with a .sdp file were untested.

This patch adjusts the stream creation logic to allow for .sdp file
generation, and renames the now more generic functions from rtsp... to
rtp...

New tests are added that create an RTP stream and generate an .sdp file.
QMediaPlayer loads the .sdp file, which should bring up a ResouceError
if an environment variable allowing it has not been set. If the
environment variable is set, the .sdp file should be played OK.

Task-number: QTBUG-125006
Pick-to: 6.7 6.5
Change-Id: I2d518979c33167aa377251af5eabb7827c978593
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>