Add HasHdrContent to QMediaMetaData::Key enumeration
The HasHdrContent metadata key can be used to determine if a video track
potentially contains HDR video.

QMediaMetaData::value(QMediaMetaData::HasHdrContent) will return true
when the decoder reports that the video uses ColorTransfer_ST2084 or
ColorTransfer_STD_B67. These are color transfers used with Dolby Vision
videos, for example HDR video recorded with an iPhone. Note that even if
HasHdrContent is true, the video may still not utilize the extended
dynamic range.

This simplified approach is chosen because it is fast, and does not
require decoding video frames to determine the actual range being used.

Qt Multimedia still does not support proper display of HDR content, but
the new metadata key can be used to determine if the video is a SDR
video that will be presented correctly.

Supported on FFmpeg media backend only.

Fixes: QTBUG-114427
Change-Id: I7361d8c61838a66a61d18a2b9e6c8d75a6bdbdb8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim@klingt.org>