Initial version of media player video format support tests
This test is intended verify format support for the media player. This
initial test covers some container formats and pixel formats, but is not
complete in any way. Only FFmpeg backend is tested.

Next step is to expand test for more container formats, pixel formats,
and also extend testing for other encoding formats.

Pick-to: 6.7 6.5
Change-Id: I1fc4a62070f7582695ceb0c51ef13e1c142c9e08
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>