ALSA: fix surround device detection
`QAlsaAudioDeviceInfo::checkSurround` did not check if a specific device
is a surround device, but if any device is a surround device. This is
both wrong and inefficient. Instead it is better to check the device
name directly.

Pick-to: 6.5
Change-Id: Id8dbcd5831d27cf5781a28c815a675cd501d4172
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit 8ec2ec24f81080b15c6061d2bd5af9478178127a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>