GStreamer: QMediaPlayer - fix output element states
When setting output elements of pipelines which have not yet been
started, we may end up in a READY->PAUSED transition forever, which can
cause getState timeouts.
Instead of setting the state of added elements to PAUSED, we sync the
state with the pipeline state. This fixes some getState timeouts when
the player gets an audio output set, but never actually plays.

Pick-to: 6.5 6.7
Change-Id: I502f3e2dc0bb67b780a37305918568e26d0a266f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>