Tests: QMediaPlayer - relax finiteLoops test
`tst_QMediaPlayerBackend::finiteLoops` validates that an expected number
of `position` updates is sent. This seems to be imprecise for the
GStreamer backend, where position updates are driven by a timer querying
the underlying GstPipeline.
So it's better to only validate it for the FFMPEG backend.

Pick-to: 6.5
Change-Id: Id998afd5ac98fad31ad38d8797d994e6445ab59a
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 621773cbaef5e0da22445eebfc88ce09ef7ae547)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>